Various fixes, Added styles, Multi Image support.
Fixed the resolutions, resolutions should work now.
Added styles, and a styles selection box, users can now select a preferred style which will greatly influence the prompt.
The generator now generates 6 images at a time, and displays all 6 images in mini previews on the canvas. if a users clicks on any of the images it will enlarge the clicked image in their default image viewer.

Known issues :
none yet reported.

This is a stand alone script. it does not need any of the other python files. it can run on its own. I suggest you guys remove the other files and just continue to expand the features on this script since it has a nice gui layout, not perfect, but nice.

users can change the tkinter theme in the gui if they prefer a different theme.